pyproject.toml
[tool.poetry]
name = "presentpy"
version = "0.1.0"
description = "Create presentations from Jupyter Notebooks"
authors = ["Antonio Feregrino <antonio.feregrino@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
nbconvert = "^6.5.0"
mistletoe = "^0.8.2"
python-pptx = "^0.6.21"
Pygments = "^2.12.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
jupyterlab = "^3.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
